refactor: split non-API admin settings into a new Features tab
The API Settings tab was mixing external-API config (VoxHub, LLM,
SMTP, OIDC keys) with feature toggles that have nothing to do with
APIs: Audio Storage, Voice Fingerprinting, and Groups. Moved all
three into a new Features tab using the existing feature-toggle
pattern, leaving API Settings focused solely on external integrations.
